requirements.txt
Flask>=3.1
Flask-Login>=0.6
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.1
openpyxl>=3.1
reportlab>=4.4
psycopg[binary]>=3.2
pytest>=8.0
